update: create a separate Component 'CameraFeed' for better scalability
